Allows empty string to pass when an item is not required
More of a discussion point than a final commit, since you'd probably want a config option or something of that nature since this changes behaviour.

I have a bunch of fields that when a form is posted the empty text fields get posted as empty strings. They aren't required, but if content in them exists it needs to be validated. Let me know if you use a different / better way of handling this.

ps. thanks for this module! out of the handful i played with, it was the nearest to exactly how i wanted things to work.